+++ src/directives/ngSelectEdit/ngSelectEdit.directive.ts
@@ -1,8 +1,9 @@
 import {Directive, ValueProvider} from "@angular/core";
 
-import {NORMAL_STATE_TYPE, LIVE_SEARCH_TYPE} from "../../misc/types";
+import {NORMAL_STATE_TYPE, LIVE_SEARCH_TYPE, POPUP_TYPE} from "../../misc/types";
 import {EditNormalStateComponent} from "../../plugins/normalState/edit/editNormalState.component";
 import {EditLiveSearchComponent} from "../../plugins/liveSearch/edit/editLiveSearch.component";
+import {EditPopupComponent} from "../../plugins/popup/edit/editPopup.component";
 
 /**
  * Directive that applies options for NgSelect which enable usage of NgSelect edit (jira like) style
@@ -21,7 +22,12 @@ import {EditLiveSearchComponent} from "../../plugins/liveSearch/edit/editLiveSea
         {
             provide: LIVE_SEARCH_TYPE,
             useValue: EditLiveSearchComponent
-        }
+        },
+        <ValueProvider>
+        {
+            provide: POPUP_TYPE,
+            useValue: EditPopupComponent
+        },
     ]
 })
 export class NgSelectEditDirective