+++ src/plugins/popup/component.ts
@@ -1,1 +1,2 @@
 export * from './basic/basicPopup.component';
+export * from './edit/editPopup.component';