+++ src/plugins/popup/index.ts
@@ -1,3 +1,4 @@
 export * from './popup.interface';
 
 export * from './basic/basicPopup.interface';
+export * from './edit/editPopup.interface';