+++ changelog.md
@@ -5,6 +5,7 @@
 - added new module `NgSelectEditModule` which contains components for displaying select with editation and tags
 - added new component `EditLiveSearchComponent` which allows is used within `EditNormalStateComponent` for editing and displaying selected value
 - added new component `EditNormalStateComponent` which uses `EditLiveSearchComponent` for editing and displaying selected value
+- added new component `EditPopupComponent` which is used for editing and displaying selected value
 - added new directive `NgSelectEditDirective` which configures `NgSelect` to use edit plugin components
 
 ## Version 6.2.0